Implement tests for contents of Docker images
Use https://github.com/aelsabbahy/goss as a framework for validating
the state of a Linux system. The built in wrapper dgoss makes it easy to
use this for docker images.

The tests check that we have:

- A php-fpm process
- Something which is listening on the expected port (we assume this is
php-fpm)
- Working Composer and Drush commands

Run the test on Circle CI for each of the PHP versions we support.
See https://circleci.com/blog/testing-docker-images-with-circleci-and-goss/