Use CGR for handling global Composer packages
We currently use a Drush specific directory for the Drush package. This
makes it tricky to add new packages as well as updating the version of
drush and other existing packages.

Use https://github.com/consolidation/cgr to replace this. It maintains
qualities regarding separation of dependencies per application while
making it easier to maintain tools through a single point of entry:
cgr.