Load built Docker images before scanning them
Before it seem like Anchore would actually load the built image
from the registry before scanning it. Now we load the image built
as a part of the workflow locally. This then allows Anchore to load
it when doing the scan.

This fails when adding PHP 8.1 - and other new tags - because these
do not exist in an external registry yet.